+++ src/views/NftOwner/index.tsx
@@ -1,17 +1,13 @@
 import { Link, } from "react-router-dom";
-import { animals } from "../../assets/utils/dummy";
+import { animals } from "../../utils/dummy";
 
 interface nftMetaData {
   name: any,
   image: any,
   tokenId: any,
-  // owner: any,
 }
 
 function NftOwner() {
-  // const { ownerAddress }:any = useParams();
-
-  // for now, define collection as a array with the json elements provided above
 
   return (
     <>
@@ -25,14 +21,12 @@ function NftOwner() {
           <div className="collection row">
             { 
               animals.map(({ name, image, tokenId, }:nftMetaData) => {
-                // if (owner === ownerAddress) {
                   return <Link key={tokenId} to={`/nftroom/`} className="col-4 yellow">
                             <div className="nft-card">
                               <img src={image} alt=""/>
                               <p>{name}</p>
                             </div>
                           </Link>
-                // }
               })
             }
           </div>
